requirements.txt
streamlit
pillow
tensorflow==2.15.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6